Rewrite common root finder in dep_path().
Rewrote the common root finder in dep_path() of gene_relations.py.

The original convoluted loop was confusing because it used a for loop in
ascending order, but checked the elements in descending order. Basically the
function finds the last common element (when searching from the right) of two
differently sized lists.