Clarify word sequence features.
Clarify the expression of word sequence features. There seems to be one if
statement which always will be executed because it is trying to compare a Word
object directly with a string. This is of course an unintentional bug on the
part of Mallory because it meant to check word.word and not the word object
directly.