Simplify words between mentions and 3-gram features.
Mallory created an unintentional bug. She should have checked for whether the third
word in the 3-gram has a comma, and not whether the entire ws array has a comma.

This bug has been intentionally replicated since I am only refactoring to get
the same output as the original program.